Refactor address logic level todo items with moderate risk in editor actions and event handling
